core-bridge: make the generator a plain library, exported without a build tag
The mobile clients no longer bind this package on its own. It is imported by the
new ../mobile wrapper, which binds it together with libbox in a single gomobile
pass, so the exported surface has to compile on every host rather than only under
the ios||android tag.

Promote generateConfig/orderNodes to GenerateConfig/OrderNodes, add a thin
ImportSubscription over the unexported fetch seam the tests inject into, and move
Version plus the ABI constant here. Drop bridge.go (the tagged surface) and
tools.go (the gomobile pin moves into mobile/).
